7. Разделяй и властвуй: еализовал подписку внутри компонента loadMoreButton и применил её
