+++ src/const.js
@@ -25,5 +25,9 @@ const RenderPosition = {
   AFTER_BEGIN: `afterbegin`,
   BEFORE_END: `beforeend`
 };
+const Event = {
+  MOUSE_DOWN: `mousedown`,
+  KEY_DOWN: `keydown`,
+};
 
-export {PROPOSALS, ACTORS, MAX_DAYS, SECONDS_IN_MINUTE, RenderPosition};
+export {PROPOSALS, ACTORS, MAX_DAYS, SECONDS_IN_MINUTE, RenderPosition, Event};
